Add files via upload
We created this branch after the deadline because of a misunderstanding about the final place where our files should be. Initially and before the deadline, I sent our project to Roger Melko in Slack. And in the morning I realized that we need to upload our files to a new branch.